refactor: restore original pages and port Phase-2 features
