+++ src/index.css
@@ -10,7 +10,7 @@
   --neon-blue: #00b3ff;
   --neon-green: #00ff8f;
   --neon-yellow: #ffe600;
-  --accent: #00e0ff;
+  --accent: #3ff;
 }
 
 body {
